+++ finance/token-swap/anchor/programs/token-swap/src/constants.rs
@@ -14,7 +14,4 @@ pub const BASIS_POINTS_DIVISOR: u64 = 10_000;
 pub const CONFIG_SEED: &[u8] = b"config";
 
 #[constant]
-pub const AUTHORITY_SEED: &[u8] = b"authority";
-
-#[constant]
 pub const LIQUIDITY_SEED: &[u8] = b"liquidity";